Module kicad_files::common

Common Syntax

This module defines all syntax that is shared across the symbol library, footprint library, schematic, board, and work sheet file formats.

Structs

Effects

Font

Justify

Paper

Point

PointList

Position

Size

TitleBlock

Enums

JustifyHoriz

JustifyVert

PaperSize