Module kicad_files::schematic

Schematic File Format

This module defines all the syntax used in the schematic file format that is not shared with symbol libraries.

Structs

Bus

BusEntry

Junction

Label

LibSymbols

NoConnect

Polyline

Schematic

Text

Version

Wire

Enums

SchematicContent