Module kicad_files::symbol

Schematic and Symbol Library Common Syntax

This module defines all syntax that is shared across the symbol library and schematic file formats.

Structs

Arc

Circle

Curve

Fill

InnerSymbol

Pin

PinAlternate

PinName

PinNames

PinNumber

PolyLine

Property

PropertyPosition

This is basically Position but for some reason, KiCAD requires the angle to be always present, whereas usually it is optional.

Rectangle

Stroke

Symbol

Text

Enums

FillType

PinElectricalType

PinGraphicalStyle

StrokeType

SymbolContent